+++ SOEStatus/ServerViewController.m
@@ -47,7 +47,7 @@
 }
 
 - (void)loadGame {
-    [SOEStatusAPI getGameStatus:self.gameId completion:^(PLRestful *api, id object, int status, NSError *error) {
+    [SOEStatusAPI getGameStatus:self.gameId completion:^(PLRestful *api, id object, NSInteger status, NSError *error) {
         if (error) {
             [PRPAlertView showWithTitle:@"Error" message:[error localizedDescription] buttonTitle:@"Continue"];
         } else {
